Refactor db as relationalDB
