+++ scripts/SavedTrips.js
@@ -3,8 +3,6 @@ import {
     getTrips,
     // getBizarreries,
     // getAllParks,
-    getBizarreryById,
-    getEateryById,
     getParkById,
     getBizarreriesByIdArray,
     getEateriesByIdArray,
@@ -16,13 +14,14 @@ import {
     directionLiteral,
 } from "./directions/DirectionDataManager.js";
 import { getEventsByParkCode } from "./parks/ParkDataManager.js";
-import { printDirectionsModal, printEventsModal, printEventsModalNone, printDirectionsModalNone } from "./modal.js";
-
+import {
+    printDirectionsModal,
+    printEventsModal,
+    printEventsModalNone,
+    printDirectionsModalNone,
+} from "./cards/modals.js";
 
 export const updateSavedTrips = () => {
-    let attractionIdToNameDictionary = {};
-    let state;
-
     const savedTripsELem = document.querySelector(".saved-trips__cards");
     savedTripsELem.innerHTML = `<div class="saved-trips__header">
                                     <h2>Saved Trips<h2>
@@ -30,11 +29,15 @@ export const updateSavedTrips = () => {
                                 <div class="saved-tips-cards-container">`;
 
     getTrips().then((tripObjs) => {
+        console.log("receiving GET:", tripObjs);
+
         // loop through trips saved in DB
         for (const tripObj of tripObjs) {
             let tripDetails = {
-                id: null,
+                id: tripObj.id,
                 parkName: null,
+                bizIdArray: [],
+                eatIdArray: [],
                 bizName: [],
                 eatName: [],
             };
@@ -42,35 +45,84 @@ export const updateSavedTrips = () => {
 
             // Make a Fetch to Bizs Eats and Parks by Id and save the name under trip details
             // console.log(tripObj.bazararieIds);
-            getBizarreriesByIdArray(tripObj.bazararieIds)
-                .then((bizObjs) => {
-                    for (const bizObj of bizObjs) {
-                        tripDetails.bizName.push(bizObj.name);
+            const joinQuery = `http://localhost:8088/joins/?tripId=${tripDetails.id}`;
+
+            fetch(joinQuery)
+                .then((response) => response.json())
+                .then((joinObjs) => {
+                    for (const joinObj of joinObjs) {
+                        if (joinObj.bizId != undefined) {
+                            tripDetails.bizIdArray.push(joinObj.bizId);
+                        }
+                        if (joinObj.eatId != undefined) {
+                            tripDetails.eatIdArray.push(joinObj.eatId);
+                        }
                     }
                 })
                 .then(() => {
-                    getEateriesByIdArray(tripObj.eateryIds)
-                        .then((eatObjs) => {
-                            for (const eatObj of eatObjs) {
-                                tripDetails.eatName.push(eatObj.businessName);
+                    // console.log(tripDetails);
+                    // console.log(
+                    //     "tripDetails.bizIdArray",
+                    //     tripDetails.bizIdArray
+                    // );
+
+                    if (tripDetails.bizIdArray.length != 0) {
+                        debugger;
+                        console.log(
+                            "tripDetails.bizIdArray",
+                            tripDetails.bizIdArray
+                        );
+
+                        getBizarreriesByIdArray(tripDetails.bizIdArray).then(
+                            (bizObjs) => {
+                                if (tripDetails.bizIdArray != []) {
+                                    // console.log(
+                                    //     "tripDetails.bizIdArray",
+                                    //     tripDetails.bizIdArray,
+                                    //     (tripDetails.bizIdArray = [])
+                                    // );
+
+                                    for (const bizObj of bizObjs) {
+                                        console.log(bizObj.name);
+                                        tripDetails.bizName.push(bizObj.name);
+                                    }
+                                }
                             }
+                        );
+                    }
+                    if (tripDetails.eatIdArray.length != 0) {
+                        getEateriesByIdArray(tripDetails.eatIdArray).then(
+                            (eatObjs) => {
+                                // console.log(
+                                //     "tripDetails.eatIdArray",
+                                //     tripDetails.eatIdArray
+                                // );
+
+                                if (tripDetails.eatIdArray != []) {
+                                    for (const eatObj of eatObjs) {
+                                        tripDetails.eatName.push(
+                                            eatObj.businessName
+                                        );
+                                    }
+                                }
+                            }
+                        );
+                    }
+                })
+                .then(() => {
+                    getParkById(tripObj.parkId)
+                        .then((park) => {
+                            tripDetails.parkName = park.fullName;
                         })
                         .then(() => {
-                            getParkById(tripObj.parkId)
-                                .then((park) => {
-                                    tripDetails.parkName = park.fullName;
-                                })
-                                .then(() => {
-                                    // when all the information is received inject the saved trip card into DOM eith the details containing names
-                                    savedTripsELem.innerHTML +=
-                                        savedTripCardDetails(
-                                            tripDetails,
-                                            tripObj.directionId
-                                        );
-                                    ///
-                                    directionsFunc(tripObj.directionId);
-                                    eventFunc(tripObj.directionId);
-                                });
+                            // when all the information is received inject the saved trip card into DOM eith the details containing names
+                            savedTripsELem.innerHTML += savedTripCardDetails(
+                                tripDetails,
+                                tripObj.directionId
+                            );
+                            // add direction and event buttons
+                            directionsFunc(tripObj.directionId);
+                            eventFunc(tripObj.directionId);
                         });
                 });
         }
@@ -78,6 +130,9 @@ export const updateSavedTrips = () => {
 };
 
 const modal = document.getElementById("modal");
+
+let currentLat;
+let currentLong;
 
 //function used for gaining permission for location
 (function () {
@@ -91,9 +146,6 @@ const modal = document.getElementById("modal");
         }
     );
 })();
-
-let currentLat;
-let currentLong;
 
 const directionsFunc = (input) => {
     //query slectors for directions button and directions fill
@@ -122,17 +174,23 @@ const directionsFunc = (input) => {
                     ).then(function (event) {
                         if (event.paths == undefined) {
                             modal.innerHTML = printDirectionsModalNone();
-                            modal.style.display = "block"
+                            modal.style.display = "block";
                         } else {
                             modal.innerHTML = printDirectionsModal();
-                            const modalDirectionsList = document.querySelector(".directions-list");
+                            const modalDirectionsList =
+                                document.querySelector(".directions-list");
                             for (
                                 let count = 0;
                                 count < event.paths[0].instructions.length;
                                 count++
                             ) {
-                                console.log(event.paths[0].instructions[count].text);
-                                modalDirectionsList.innerHTML += directionLiteral(event.paths[0].instructions[count].text);
+                                console.log(
+                                    event.paths[0].instructions[count].text
+                                );
+                                modalDirectionsList.innerHTML +=
+                                    directionLiteral(
+                                        event.paths[0].instructions[count].text
+                                    );
                                 modal.style.display = "block";
                             }
                         }
@@ -143,9 +201,6 @@ const directionsFunc = (input) => {
     });
 };
 
-
-
-
 //Events Function
 export const eventFunc = (input) => {
     //obtain container for eventListener
@@ -153,10 +208,7 @@ export const eventFunc = (input) => {
 
     //add eventListener for on clic;
     directionElement.addEventListener("click", (event) => {
-
-        //if statement to decide which card the user selected
         if (event.target.id == `events-btn--${input}`) {
-
             //fetch calls to gather the trip selected--then gather the parkId of selected trip and pass that parkId to the getParks to then get activities from that specific park
             const getTrip = getSingleTripByDirectionId(input).then((taco) => {
                 getParkById(taco[0].parkId).then((parkEvent) => {
@@ -168,14 +220,18 @@ export const eventFunc = (input) => {
                             modal.style.display = "block";
                         } else if (event.length == 1) {
                             for (let count = 0; count < event.length; count++) {
-                                modal.innerHTML = printEventsModal(event[count]);
+                                modal.innerHTML = printEventsModal(
+                                    event[count]
+                                );
                                 modal.style.display = "block";
-                            };
+                            }
                         } else {
                             for (let count = 0; count < 2; count++) {
-                                modal.innerHTML = printEventsModal(event[count]);
+                                modal.innerHTML = printEventsModal(
+                                    event[count]
+                                );
                                 modal.style.display = "block";
-                            };
+                            }
                         }
                     });
                 });
