+++ scripts/eateries/EateryDataManager.js
@@ -2,28 +2,25 @@ import { updateActiveTrip } from "../main.js";
 //Export for Eateries
 export const getEateries = () => {
     return fetch("http://holidayroad.nss.team/eateries")
-    .then(response => response.json())
-    .then(parsedResponse => {
-        return parsedResponse;
-    })
-}
+        .then((response) => response.json())
+        .then((parsedResponse) => {
+            return parsedResponse;
+        });
+};
 
 export const getEatNameById = (eatId) => {
     return fetch(`http://holidayroad.nss.team/eateries/?id=${eatId}`)
-    .then(response => response.json())
-    .then(parsedResponse => {
-        return parsedResponse;
-    })
-}
+        .then((response) => response.json())
+        .then((parsedResponse) => {
+            return parsedResponse;
+        });
+};
 
 export const onEatChanged = (eat) => {
     const eatNamePreview = document.getElementById("eatPreview");
 
-    const getEatVar = getEatNameById(eat).then(eatName => {
+    const getEatVar = getEatNameById(eat).then((eatName) => {
         eatNamePreview.innerHTML = eatName[0].businessName;
-    })
+    });
     updateActiveTrip("eateryIds", eat);
-}
-
-
-
+};
